requirements.txt
Django>=4.0.3,<4.1
django-cors-headers>=3.11.0,<3.12
django-filter>=21.1,<21.2
djangorestframework>=3.13.1,<3.14
playsound>=1.3.0,<1.4
pytz>=2022.1,<2022.2
sqlparse>=0.4.2,<0.5
tzdata>=2022.1,<2022.2
psycopg2>=2.9.3,<3.0
uWSGI>=2.0.19.1,<2.1
asgiref>=3.5.0,<3.6
Pillow>=9.1.1,<9.2